+++ crates/nostr-database/src/events/mod.rs
@@ -119,27 +119,6 @@ pub trait NostrEventsDatabase: fmt::Debug + Send + Sync {
         timestamp: &'a Timestamp,
     ) -> BoxedFuture<'a, Result<bool, DatabaseError>>;
 
-    /// Set [`EventId`] as seen by relay
-    ///
-    /// Useful for NIP65 (aka gossip)
-    #[deprecated(since = "0.39.0")]
-    fn event_id_seen(
-        &self,
-        event_id: EventId,
-        relay_url: RelayUrl,
-    ) -> BoxedFuture<Result<(), DatabaseError>>;
-
-    /// Get list of relays that have seen the [`EventId`]
-    #[deprecated(
-        since = "0.39.0",
-        note = "For now this method hasn't a replacement and maybe will never have one. \
-        You can keep track of seen IDs by looking at the Relay, RelayPool or Client notifications."
-    )]
-    fn event_seen_on_relays<'a>(
-        &'a self,
-        event_id: &'a EventId,
-    ) -> BoxedFuture<'a, Result<Option<HashSet<RelayUrl>>, DatabaseError>>;
-
     /// Get [`Event`] by [`EventId`]
     fn event_by_id<'a>(
         &'a self,