+++ crates/nostr-lmdb/src/lib.rs
@@ -10,7 +10,6 @@
 #![warn(rustdoc::bare_urls)]
 #![allow(clippy::mutable_key_type)]
 
-use std::collections::HashSet;
 use std::path::Path;
 
 use nostr_database::prelude::*;
@@ -100,21 +99,6 @@ impl NostrEventsDatabase for NostrLMDB {
                 Ok(false)
             }
         })
-    }
-
-    fn event_id_seen(
-        &self,
-        _event_id: EventId,
-        _relay_url: RelayUrl,
-    ) -> BoxedFuture<Result<(), DatabaseError>> {
-        Box::pin(async move { Ok(()) })
-    }
-
-    fn event_seen_on_relays<'a>(
-        &'a self,
-        _event_id: &'a EventId,
-    ) -> BoxedFuture<'a, Result<Option<HashSet<RelayUrl>>, DatabaseError>> {
-        Box::pin(async move { Err(DatabaseError::NotSupported) })
     }
 
     fn event_by_id<'a>(